seperate module distribution.

+++ js/ydn/db/base/db.js
@@ -19,9 +19,17 @@
  */
 
 goog.provide('ydn.db');
-goog.require('ydn.db.con.IndexedDb');
-goog.require('ydn.db.con.WebSql');
 goog.require('ydn.db.utils');
+if (!ydn.db.base.NO_IDB) {
+  goog.require('ydn.db.con.IndexedDb');
+}
+if (!ydn.db.base.NO_WEBSQL) {
+  goog.require('ydn.db.con.WebSql');
+}
+if (!ydn.db.base.NO_SIMPLE) {
+  goog.require('ydn.db.con.LocalStorage');
+  goog.require('ydn.db.con.SessionStorage');
+}
 
 /**
  *
@@ -40,20 +48,20 @@ ydn.db.deleteDatabase = function(db_name, type) {
   // http://www.w3.org/TR/IndexedDB/#widl-IDBFactory-deleteDatabase-IDBOpenDBRequest-DOMString-name
 
   // some IndexedDB API do not support deleting database.
-  if (ydn.db.con.IndexedDb.isSupported() && (!type ||
+  if (!ydn.db.base.NO_IDB && ydn.db.con.IndexedDb.isSupported() && (!type ||
     type == ydn.db.con.IndexedDb.TYPE) &&
     ydn.db.con.IndexedDb.indexedDb &&
       ('deleteDatabase' in ydn.db.con.IndexedDb.indexedDb)) {
     ydn.db.con.IndexedDb.indexedDb.deleteDatabase(db_name);
   }
-  if (ydn.db.con.WebSql.isSupported() && (!type ||
+  if (!ydn.db.base.NO_WEBSQL && ydn.db.con.WebSql.isSupported() && (!type ||
     type == ydn.db.con.WebSql.TYPE)) {
     ydn.db.con.WebSql.deleteDatabase(db_name);
   }
-  if (!type || type == ydn.db.con.LocalStorage.TYPE) {
+  if (!ydn.db.base.NO_SIMPLE && (!type || type == ydn.db.con.LocalStorage.TYPE)) {
     ydn.db.con.LocalStorage.deleteDatabase(db_name);
   }
-  if (!type || type == ydn.db.con.SessionStorage.TYPE) {
+  if (!ydn.db.base.NO_SIMPLE && (!type || type == ydn.db.con.SessionStorage.TYPE)) {
     ydn.db.con.SessionStorage.deleteDatabase(db_name);
   }
 };
@@ -65,7 +73,7 @@ ydn.db.deleteDatabase = function(db_name, type) {
  * greater than the second, -1 if the first is less than the second, and 0 if
  * the first is equal to the second.
  */
-ydn.db.cmp = (ydn.db.con.IndexedDb.indexedDb &&
+ydn.db.cmp = (!ydn.db.base.NO_IDB && ydn.db.con.IndexedDb.indexedDb &&
   ydn.db.con.IndexedDb.indexedDb.cmp) ?
     goog.bind(ydn.db.con.IndexedDb.indexedDb.cmp,
       ydn.db.con.IndexedDb.indexedDb) : ydn.db.utils.cmp;
